pyproject.toml
[project]
name = "pdfspy"
version = "0.1.0"
description = "High-performance PDF data extraction with chemical NER and Donut integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pdfplumber>=0.11.0",
    "rapidfuzz>=3.6.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "pdf2image>=1.17.0",
    "joblib>=1.3.0",
    "pillow>=10.0.0",
    "datasets>=2.14.0",
    "accelerate>=0.24.0",
    "pandas>=2.1.0",
]

[tool.pytest.ini_options]
markers = [
    "performance: marks tests as performance tests (deselect with '-m \"not performance\"')",
]

[tool.pyrefly]
project-includes = ["**/*"]
project-excludes = ["**/*venv/**/*"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
